[F]MOD: 美化了login页面，美化并合并了充值界面
