+++ src/main/resources/view/fxml/CustomPage.fxml
@@ -5,24 +5,24 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.CustomController">
+<AnchorPane prefHeight="300.0" prefWidth="500.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.CustomController">
    <children>
-      <Label layoutX="111.0" layoutY="80.0" text="Please enter the amount of topping up:">
+      <Label layoutX="59.0" layoutY="47.0" text="Please enter the amount of topping up:">
          <font>
             <Font size="22.0" />
          </font>
       </Label>
-      <TextField fx:id="amountTextField" layoutX="231.0" layoutY="176.0" prefHeight="48.0" prefWidth="141.0">
+      <TextField fx:id="amountTextField" layoutX="191.0" layoutY="120.0" prefHeight="48.0" prefWidth="141.0">
          <font>
             <Font size="22.0" />
          </font>
       </TextField>
-      <Button layoutX="247.0" layoutY="288.0" mnemonicParsing="false" onMouseClicked="#handleConfirm" text="Confirm">
+      <Button layoutX="196.0" layoutY="208.0" mnemonicParsing="false" onMouseClicked="#handleConfirm" text="Confirm">
          <font>
             <Font size="22.0" />
          </font>
       </Button>
-      <Label layoutX="209.0" layoutY="184.0" text="\$">
+      <Label layoutX="169.0" layoutY="128.0" text="\$">
          <font>
             <Font size="22.0" />
          </font>
