+++ src/main/resources/view/fxml/LoginPage.fxml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import javafx.scene.shape.*?>
 <?import javafx.scene.paint.*?>
 <?import javafx.scene.text.*?>
 <?import java.lang.*?>
@@ -8,31 +9,66 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<AnchorPane prefHeight="500.0" prefWidth="700.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.LoginController">
+<AnchorPane prefHeight="350.0" prefWidth="500.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.LoginController">
     <children>
-        <Separator layoutY="95.0" prefWidth="200.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" />
-        <Label layoutX="35.0" layoutY="23.0" text="MyAPP" textFill="#70796f">
-         <font>
-            <Font name="System Bold" size="36.0" />
-         </font></Label>
-        <Button fx:id="LoginButton" layoutX="175.0" layoutY="368.0" mnemonicParsing="false" onMouseClicked="#login" prefHeight="50.0" prefWidth="350.0" text="Login">
+      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="WHITE" height="224.0" layoutY="126.0" stroke="WHITE" strokeType="INSIDE" width="500.0" />
+        <Button fx:id="LoginButton" alignment="CENTER" layoutX="110.0" layoutY="266.0" mnemonicParsing="false" onMouseClicked="#login" prefHeight="40.0" prefWidth="280.0" style="-fx-background-radius: 5;" text="Login" textFill="WHITE" AnchorPane.bottomAnchor="35.0" AnchorPane.leftAnchor="110.0" AnchorPane.rightAnchor="110.0">
          <font>
             <Font size="24.0" />
-         </font></Button>
-        <TextField fx:id="account" layoutX="300.0" layoutY="205.0" />
-        <ImageView fx:id="imageView" fitHeight="100.0" fitWidth="100.0" layoutX="300.0" layoutY="48.0" pickOnBounds="true" scaleZ="20.0">
-            <image>
-                <Image url="@../images/coach.jpg" />
-            </image>
-            <viewport>
-                <Rectangle2D />
-            </viewport></ImageView>
-        <Button layoutX="35.0" layoutY="438.0" mnemonicParsing="false" onAction="#showSignUpPage" style="-fx-background-color: transparent;" text="Sign Up" textFill="#bc1515" textOverrun="CLIP" underline="true">
+         </font>
+         <padding>
+            <Insets top="1.0" />
+         </padding></Button>
+        <TextField fx:id="account" alignment="BOTTOM_LEFT" layoutX="186.0" layoutY="195.0" maxWidth="200.0" style="-fx-text-box-border: transparent; -fx-background-color: -fx-control-inner-background;" AnchorPane.bottomAnchor="150.0" AnchorPane.rightAnchor="110.0">
          <font>
-            <Font name="System Bold Italic" size="18.0" />
+            <Font size="22.0" />
+         </font>
+         <padding>
+            <Insets top="1.0" />
+         </padding></TextField>
+        <Button layoutX="13.0" layoutY="347.0" mnemonicParsing="false" onAction="#showSignUpPage" style="-fx-background-color: transparent;" text="Sign Up" textFill="#b5b0b0" textOverrun="CLIP" AnchorPane.bottomAnchor="5.0" AnchorPane.leftAnchor="5.0">
+         <font>
+            <Font size="20.0" />
          </font></Button>
-        <Label layoutX="187.0" layoutY="211.0" text="Account:" />
-        <Label layoutX="184.0" layoutY="299.0" text="Password:" />
-      <PasswordField fx:id="password" layoutX="300.0" layoutY="293.0" />
+        <Label layoutX="97.0" layoutY="168.0" text="Account:" textFill="#04b9f9" AnchorPane.bottomAnchor="146.0" AnchorPane.rightAnchor="315.0">
+         <font>
+            <Font name="System Italic" size="22.0" />
+         </font></Label>
+        <Label layoutX="75.0" layoutY="218.0" text="Password:" textFill="#04b9f9" AnchorPane.bottomAnchor="96.0" AnchorPane.rightAnchor="315.0">
+         <font>
+            <Font name="System Italic" size="22.0" />
+         </font></Label>
+      <PasswordField fx:id="password" alignment="BOTTOM_LEFT" layoutX="166.0" layoutY="250.0" maxWidth="200.0" style="-fx-text-box-border: transparent; -fx-background-color: -fx-control-inner-background;" AnchorPane.bottomAnchor="100.0" AnchorPane.rightAnchor="110.0">
+         <font>
+            <Font size="22.0" />
+         </font>
+         <padding>
+            <Insets top="1.0" />
+         </padding></PasswordField>
+      <Rectangle fx:id="rectangle" arcHeight="5.0" arcWidth="5.0" height="125.0" layoutX="30.0" layoutY="14.0" strokeType="INSIDE" width="500.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+         <stroke>
+            <LinearGradient endX="1.0" endY="1.0">
+               <stops>
+                  <Stop color="#2f71f5" />
+                  <Stop color="#7c2ff5" offset="0.2914285714285714" />
+                  <Stop color="#17bffc" offset="0.6028571428571429" />
+                  <Stop color="#1f70ff" offset="1.0" />
+               </stops>
+            </LinearGradient>
+         </stroke>
+         <fill>
+            <LinearGradient endX="1.0" endY="1.0">
+               <stops>
+                  <Stop color="#7c2ff5" />
+                  <Stop color="#7c2ff5" offset="0.00625" />
+                  <Stop color="#2f71f5" offset="0.30000000000000004" />
+                  <Stop color="#17bffc" offset="0.6028571428571429" />
+                  <Stop color="#1f70ff" offset="1.0" />
+               </stops>
+            </LinearGradient>
+         </fill>
+      </Rectangle>
+      <Line endX="100.0" layoutX="190.0" layoutY="98.0" startX="-100.0" stroke="#a8a8a8" AnchorPane.bottomAnchor="145.0" AnchorPane.rightAnchor="110.0" />
+      <Line endX="100.0" layoutX="250.0" layoutY="79.0" startX="-100.0" stroke="#a8a8a8" AnchorPane.bottomAnchor="95.0" AnchorPane.rightAnchor="110.0" />
     </children>
 </AnchorPane>
