--- src/main/resources/view/fxml/TopUp.fxml
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import javafx.scene.text.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<AnchorPane prefHeight="675.0" prefWidth="1000.0" style="-fx-background-color: white;" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.TopUpController">
-   <children>
-      <Label layoutX="27.0" layoutY="37.0" prefHeight="90.0" prefWidth="354.0" text="Top up for your account:">
-         <font>
-            <Font size="31.0" />
-         </font>
-      </Label>
-      <Button fx:id="confirmButton" layoutX="434.0" layoutY="513.0" mnemonicParsing="false" onMouseClicked="#handleConfirm" prefHeight="64.0" prefWidth="132.0" text="Confirm">
-         <font>
-            <Font size="22.0" />
-         </font>
-      </Button>
-      <Button fx:id="backButton" layoutX="834.0" layoutY="58.0" mnemonicParsing="false" onMouseClicked="#backToAccount" prefHeight="38.0" prefWidth="97.0" text="Back">
-         <font>
-            <Font size="22.0" />
-         </font>
-      </Button>
-      <Label fx:id="l1" alignment="CENTER" layoutX="64.0" layoutY="171.0" onMouseClicked="#selectAmount" prefHeight="84.0" prefWidth="174.0" style="-fx-border-color: black;" text="\$6">
-         <font>
-            <Font size="36.0" />
-         </font>
-      </Label>
-      <Label fx:id="l2" alignment="CENTER" layoutX="300.0" layoutY="171.0" onMouseClicked="#selectAmount" prefHeight="84.0" prefWidth="174.0" style="-fx-border-color: black;" text="\$18">
-         <font>
-            <Font size="36.0" />
-         </font>
-      </Label>
-      <Label fx:id="l3" alignment="CENTER" layoutX="540.0" layoutY="171.0" onMouseClicked="#selectAmount" prefHeight="84.0" prefWidth="174.0" style="-fx-border-color: black;" text="\$68">
-         <font>
-            <Font size="36.0" />
-         </font>
-      </Label>
-      <Label fx:id="l4" alignment="CENTER" layoutX="787.0" layoutY="171.0" onMouseClicked="#selectAmount" prefHeight="84.0" prefWidth="174.0" style="-fx-border-color: black;" text="\$233">
-         <font>
-            <Font size="36.0" />
-         </font>
-      </Label>
-      <Label fx:id="l5" alignment="CENTER" layoutX="64.0" layoutY="359.0" onMouseClicked="#selectAmount" prefHeight="84.0" prefWidth="174.0" style="-fx-border-color: black;" text="\$648">
-         <font>
-            <Font size="36.0" />
-         </font>
-      </Label>
-      <Label fx:id="l6" alignment="CENTER" layoutX="300.0" layoutY="359.0" onMouseClicked="#selectAmount" prefHeight="84.0" prefWidth="174.0" style="-fx-border-color: black;" text="\$998">
-         <font>
-            <Font size="36.0" />
-         </font>
-      </Label>
-      <Label fx:id="l7" alignment="CENTER" layoutX="540.0" layoutY="359.0" onMouseClicked="#selectAmount" prefHeight="84.0" prefWidth="174.0" style="-fx-border-color: black;" text="\$8848">
-         <font>
-            <Font size="36.0" />
-         </font>
-      </Label>
-      <Label fx:id="l8" alignment="CENTER" layoutX="787.0" layoutY="359.0" onMouseClicked="#selectAmount" prefHeight="84.0" prefWidth="174.0" style="-fx-border-color: black;" text="Custom">
-         <font>
-            <Font size="36.0" />
-         </font>
-      </Label>
-      <Label fx:id="amountLabel" layoutX="481.0" layoutY="38.0" text="\$0" textFill="#ff2400">
-         <font>
-            <Font name="System Bold Italic" size="60.0" />
-         </font>
-      </Label>
-   </children>
-</AnchorPane>
